feat: 생성된 게임만 필터에 뜨게 변경
