+++ src/entities/community/api/getStageGame.ts
