+++ src/entities/community/model/useGetStageGameQuery.ts
