+++ src/entities/community/ui/Modal/index.tsx
@@ -13,6 +13,7 @@ interface ModalProps {
   selectedSort?: SortType | null;
   toggleSportSelection: (sport: SportType) => void;
   toggleSortSelection?: (sort: SortType) => void;
+  categoryTypes: SportType[];
 }
 
 const Modal = ({
@@ -21,16 +22,8 @@ const Modal = ({
   selectedSort,
   toggleSportSelection,
   toggleSortSelection,
+  categoryTypes,
 }: ModalProps) => {
-  const categoryTypes: SportType[] = [
-    'VOLLEY_BALL',
-    'SOCCER',
-    'LOL',
-    'BASE_BALL',
-    'BASKET_BALL',
-    'BADMINTON',
-    'ETC',
-  ];
   const sortTypes: SortType[] = ['LATEST', 'LAST'];
 
   return (
