+++ src/views/community/ui/CommunityPage/index.tsx
@@ -1,6 +1,7 @@
 'use client';
 
 import { useParams, useSearchParams } from 'next/navigation';
+import { useCategoryTypes } from '@/entities/community/model/useCategoryTypes';
 import useSelectSort from '@/shared/model/useSelectSort';
 import useSelectSport from '@/shared/model/useSelectSport';
 import BackPageButton from '@/shared/ui/backPageButton';
@@ -15,6 +16,8 @@ const CommunityPage = () => {
 
   const { selectedSort, toggleSortSelection } = useSelectSort();
   const { selectedSport, toggleSportSelection } = useSelectSport();
+
+  const { categoryTypes } = useCategoryTypes(stageId);
 
   return (
     <div
@@ -35,13 +38,14 @@ const CommunityPage = () => {
         <div
           className={cn('w-full', 'h-full', 'flex', 'flex-col', 'gap-[2rem]')}
         >
-          <BackPageButton />
+          <BackPageButton type="push" path={`/${stageId}`} />
           <CommunityToolbar
             selectedSport={selectedSport}
             selectedSort={selectedSort}
             toggleSportSelection={toggleSportSelection}
             toggleSortSelection={toggleSortSelection}
             stageId={stageId}
+            categoryTypes={categoryTypes}
           />
           <CommunityItemContainer
             stageId={stageId}
